+++ 01-cli/src/main/kotlin/ru/spbau/bogomolov/Main.kt
@@ -1,5 +1,7 @@
 package ru.spbau.bogomolov
 
+import ru.spbau.bogomolov.executor.BashLikeExecutor
+
 fun main(args : Array<String>) {
     val appInstance = CommandLineInterpretator()
     val interpretator = BashLikeExecutor()
