+++ 01-cli/src/main/kotlin/ru/spbau/bogomolov/executor/ExecutionResult.kt
@@ -1,4 +1,4 @@
-package ru.spbau.bogomolov
+package ru.spbau.bogomolov.executor
 
 
 /**
