+++ 01-cli/src/main/kotlin/ru/spbau/bogomolov/parser/BashLikeParser.kt
@@ -1,0 +1,9 @@
+package ru.spbau.bogomolov.parser
+
+import ru.spbau.bogomolov.ast.AstNode
+
+class BashLikeParser : CommandLineParser {
+    override fun parse(string: String): AstNode {
+
+    }
+}
