+++ 01-cli/src/main/kotlin/ru/spbau/bogomolov/parser/CommandLineParser.kt
@@ -1,0 +1,10 @@
+package ru.spbau.bogomolov.parser
+
+import ru.spbau.bogomolov.ast.AstNode
+
+/**
+ * Processes a string and returns a root of AST corresponding to it.
+ */
+interface CommandLineParser {
+    fun parse(string: String): AstNode
+}
